pyproject.toml
[project]
name = "dexbooru-ai"
version = "1.0.0"
description = "A microservice to handle AI operations like vector search, RAG and basic ML on the Dexbooru image board"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp[speedups]>=3.13.3",
    "fastapi[standard]>=0.135.1",
    "google-genai>=1.0.0",
    "pika>=1.3.2",
    "pillow>=12.1.1",
    "pydantic-settings>=2.13.1",
    "qdrant-client>=1.17.1",
    "uvicorn>=0.41.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.0.0",
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
testpaths = ["test"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 200
src = ["src", "test"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.format]
quote-style = "double"
